requirements.txt
# TARS Chatbot v2 - Enhanced Architecture
# Dependencies and requirements for the project

# Core AI/ML
openai>=1.12.0
google-generativeai>=0.4.0
sentence-transformers>=2.3.0
chromadb>=0.4.22
langchain>=0.1.0
langchain-community>=0.0.20

# Voice
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.14

# Backend API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Data Processing
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# NLP
nltk>=3.8.1
tiktoken>=0.5.2

# Utilities
rich>=13.7.0
typer>=0.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Development
black>=24.1.0
isort>=5.13.0
